Add Github count on homepage
