+++ apps/frontpage/lib/get-github-count.ts
@@ -1,0 +1,10 @@
+export const githubCount = async () => {
+  const response = await fetch(
+    'https://api.github.com/repos/storybookjs/storybook'
+  );
+  const data = await response.json();
+
+  if (!response.ok) return 0;
+
+  return data.stargazers_count;
+};
